requirements.txt
Django==5.1
djangorestframework==3.15.2
elastic-transport==8.15.0
elasticsearch==8.15.0
psycopg2==2.9.9
sqlparse==0.5.1
urllib3==2.2.2